backend/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()

DATABASE_URL = os.getenv(
    "DATABASE_URL",
    "postgresql://postgres:postgres@localhost:5432/wikiquiz"
)

GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")
