backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
beautifulsoup4
requests
langchain
langchain-google-genai
pydantic
python-dotenv
